# backend/internal/data/repo/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package repo

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Frame struct {
	ID        pgtype.UUID        `json:"ID"`
	StreamID  pgtype.UUID        `json:"StreamID"`
	Sequence  int32              `json:"Sequence"`
	Payload   []byte             `json:"Payload"`
	MimeType  string             `json:"MimeType"`
	CreatedAt pgtype.Timestamptz `json:"CreatedAt"`
}

type Stream struct {
	ID              pgtype.UUID        `json:"ID"`
	Title           string             `json:"Title"`
	Description     string             `json:"Description"`
	FrameIntervalMs int32              `json:"FrameIntervalMs"`
	CreatedAt       pgtype.Timestamptz `json:"CreatedAt"`
	UpdatedAt       pgtype.Timestamptz `json:"UpdatedAt"`
}